RawServerErrorPage now respond with status 500 (#806)
