+++ src/main/java/io/kyberorg/yalsee/exception/RawLoopException.java
@@ -1,0 +1,11 @@
+package io.kyberorg.yalsee.exception;
+
+import io.kyberorg.yalsee.ui.err.raw.RawServerErrorPage;
+
+/**
+ * Exception, which not intended to be thrown, just {@link RawServerErrorPage} handles it.
+ *
+ * @since 3.11
+ */
+public class RawLoopException extends YalseeException {
+}
